Move all LevelDB operations to ldb.py
